Feat: useDropdown 훅 추가 및 적용
dropdown을 사용하는 부분에서 공통적으로 작성되는 부분을 묶어 훅으로 만들었습니다.
드롭다운을 열고 닫기 위한 state인 isDropdownOpen, setIsDropdownOpen이 있고, 인풋이 포커스 되거나 풀릴 때 드롭다운을 열고 닫는 동작을 하는 handleFocus, handleBlur 함수가 있습니다.
inputRef는 포커스 상태를 변경하는데에 사용합니다. 사용하는 input에 ref={inputRef}를 걸어 포커스 관련 동작을 사용할 수 있습니다.

+작명을 suggestive-search-input에서 product-suggestive-search-input으로 이름 변경 예정
